backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
python-dotenv==1.0.1
google-generativeai==0.4.0
PyPDF2==3.0.1
pikepdf==9.7.0
reportlab==4.4.1
pytesseract==0.3.13
